noveo_back/requirements.txt
Django==2.2
django-cors-headers==3.0.2
djangorestframework==3.10.2
passlib==1.7.1
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
